Split big main file
